requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
alembic==1.12.0
pydantic==1.8.2
SQLAlchemy==1.4.41
sqlmodel==0.0.8
dependency-injector==4.41.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9